pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "multi-sources-financial-control"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "^3.2.5"
django-choices = "^1.7.2"
python = "^3.8"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.2"
django-filter = "^2.4.0"
drf-spectacular = "0.17.3"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
factory-boy = "^3.2.0"
pylint-django = "^2.4.4"
pylint-runner = "^0.6.0"
pytest = "^5.2"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
pytest-mock = "^3.6.1"
pytest-sugar = "^0.9.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length = 88
